Removed multipointer, updated files using it, linting
