+++ pyfms/__init__.py
@@ -1,6 +1,7 @@
 from .py_data_override.py_data_override import pyDataOverride
+from .py_diag_manager.pyfms_diag_manager import pyFMS_diag_manager
 from .py_field_manager.py_field_manager import FieldTable
-from .py_horiz_interp.py_horiz_interp import HorizInterp
+from .py_horiz_interp.py_horiz_interp import pyHorizInterp
 from .py_mpp.py_mpp import pyFMS_mpp
 from .py_mpp.py_mpp_domains import (
     pyDomain,
