+++ pyfms/py_mpp/py_mpp_domains.py
@@ -6,7 +6,7 @@ from numpy.typing import NDArray
 
 from ..pyfms_utils.data_handling import (
     set_Cchar,
-    set_multipointer,
+    setarray_Cbool,
     setarray_Cint32,
     setscalar_Cbool,
     setscalar_Cint32,
@@ -70,6 +70,7 @@ class pyFMS_mpp_domains:
         xextent: Optional[NDArray] = None,
         yextent: Optional[NDArray] = None,
         maskmap: Optional[NDArray[np.bool_]] = None,
+        maskmap_shape: Optional[NDArray] = None,
         name: Optional[str] = None,
         symmetry: Optional[bool] = None,
         memory_size: Optional[NDArray] = None,
@@ -97,7 +98,8 @@ class pyFMS_mpp_domains:
         yhalo_c, yhalo_t = setscalar_Cint32(yhalo)
         xextent_p, xextent_t = setarray_Cint32(xextent)
         yextent_p, yextent_t = setarray_Cint32(yextent)
-        maskmap_p, maskmap_t = set_multipointer(arg=maskmap, num_ptr=2)
+        maskmap_p, maskmap_t = setarray_Cbool(maskmap)
+        maskmap_shape_p, maskmap_shape_t = setarray_Cint32(maskmap_shape)
         name_c, name_t = set_Cchar(name)
         symmetry_c, symmetry_t = setscalar_Cbool(symmetry)
         memory_size_p, memory_size_t = setarray_Cint32(memory_size)
@@ -124,6 +126,7 @@ class pyFMS_mpp_domains:
             xextent_t,
             yextent_t,
             maskmap_t,
+            maskmap_shape_t,
             name_t,
             symmetry_t,
             memory_size_t,
@@ -153,6 +156,7 @@ class pyFMS_mpp_domains:
             xextent_p,
             yextent_p,
             maskmap_p,
+            maskmap_shape_p,
             name_c,
             symmetry_c,
             memory_size_p,
