+++ pyfms/pyfms_utils/data_handling.py
@@ -99,108 +99,6 @@ def setarray_Cint32(
 
 
 """
-set_multipointer:
-    For converting NumPy arrays to pointer equivalent
-    This method is able to convert a Numpy array of
-    up to 5 dimensions to quintuple pointer (*****).
-
-    If an array of <= 1 or > 5 dimensions is passed
-    as an argument, the method will default to the
-    returning the passed array and the result of a
-    pass to set_ndpointer.
-
-    This method is expected to be slow, and when
-    possible avoided.
-"""
-
-
-def set_multipointer(arg: npt.NDArray, num_ptr: int) -> Tuple:
-    if arg is not None:
-        c_type = np.ctypeslib.as_ctypes_type(arg.dtype)
-    match num_ptr:
-        case 2:
-            if arg is not None:
-                arg_ptr = (ctypes.POINTER(c_type) * arg.shape[0])()
-                for i in range(arg.shape[0]):
-                    arg_ptr[i] = arg[i].ctypes.data_as(ctypes.POINTER(c_type))
-                return arg_ptr, ctypes.POINTER(ctypes.POINTER(c_type))
-            else:
-                return arg, ctypes.POINTER(ctypes.POINTER(ctypes.c_int))
-        case 3:
-            if arg is not None:
-                arg_ptr = (ctypes.POINTER(ctypes.POINTER(c_type)) * arg.shape[0])()
-                for i in range(arg.shape[0]):
-                    arg_ptr[i] = (ctypes.POINTER(c_type) * arg.shape[1])()
-                    for j in range(arg.shape[1]):
-                        arg_ptr[i][j] = arg[i][j].ctypes.data_as(ctypes.POINTER(c_type))
-                return arg_ptr, ctypes.POINTER(ctypes.POINTER(ctypes.POINTER(c_type)))
-            else:
-                return arg, ctypes.POINTER(ctypes.POINTER(ctypes.POINTER(ctypes.c_int)))
-        case 4:
-            if arg is not None:
-                arg_ptr = (
-                    ctypes.POINTER(ctypes.POINTER(ctypes.POINTER(c_type)))
-                    * arg.shape[0]
-                )()
-                for i in range(arg.shape[0]):
-                    arg_ptr[i] = (
-                        ctypes.POINTER(ctypes.POINTER(c_type)) * arg.shape[1]
-                    )()
-                    for j in range(arg.shape[1]):
-                        arg_ptr[i][j] = (ctypes.POINTER(c_type) * arg.shape[2])()
-                        for k in range(arg.shape[2]):
-                            arg_ptr[i][j][k] = arg[i][j][k].ctypes.data_as(
-                                ctypes.POINTER(c_type)
-                            )
-                return arg_ptr, ctypes.POINTER(
-                    ctypes.POINTER(ctypes.POINTER(ctypes.POINTER(c_type)))
-                )
-            else:
-                return arg, ctypes.POINTER(
-                    ctypes.POINTER(ctypes.POINTER(ctypes.POINTER(ctypes.c_int)))
-                )
-        case 5:
-            if arg is not None:
-                arg_ptr = (
-                    ctypes.POINTER(
-                        ctypes.POINTER(ctypes.POINTER(ctypes.POINTER(c_type)))
-                    )
-                    * arg.shape[0]
-                )()
-                for i in range(arg.shape[0]):
-                    arg_ptr[i] = (
-                        ctypes.POINTER(ctypes.POINTER(ctypes.POINTER(c_type)))
-                        * arg.shape[1]
-                    )()
-                    for j in range(arg.shape[1]):
-                        arg_ptr[i][j] = (
-                            ctypes.POINTER(ctypes.POINTER(c_type)) * arg.shape[2]
-                        )()
-                        for k in range(arg.shape[2]):
-                            arg_ptr[i][j][k] = (ctypes.POINTER(c_type) * arg.shape[3])()
-                            for n in range(arg.shape[3]):
-                                arg_ptr[i][j][k][n] = arg[i][j][k][n].ctypes.data_as(
-                                    ctypes.POINTER(c_type)
-                                )
-                return arg_ptr, ctypes.POINTER(
-                    ctypes.POINTER(
-                        ctypes.POINTER(ctypes.POINTER(ctypes.POINTER(c_type)))
-                    )
-                )
-            else:
-                return arg, ctypes.POINTER(
-                    ctypes.POINTER(
-                        ctypes.POINTER(ctypes.POINTER(ctypes.POINTER(ctypes.c_int)))
-                    )
-                )
-        case _:
-            if arg is not None:
-                return arg, set_ndpointer(arg)
-            else:
-                return arg, ctypes.POINTER(ctypes.c_int)
-
-
-"""
 Scalar setting methods
 """
